Add `visit/2` to Driver protocol
In order to support external libraries implementing the `Driver`
protocol, we move the `visit/2` call out of `PhoenixTest`. When it's a
`Plug.Conn` struct, we delegate to `ConnHandler.visit/2` (not the best
name, but good enough for now).
