+++ lib/phoenix_test.ex
@@ -194,14 +194,13 @@ defmodule PhoenixTest do
   For more info, see `within/3`.
   """
 
-  import Phoenix.ConnTest
   import PhoenixTest.Locators
 
+  alias PhoenixTest.ConnHandler
   alias PhoenixTest.Driver
   alias PhoenixTest.Element.Button
   alias PhoenixTest.Query
 
-  @endpoint Application.compile_env(:phoenix_test, :endpoint)
   @doc """
   Entrypoint to create a session.
 
@@ -213,25 +212,12 @@ defmodule PhoenixTest do
   LiveView or a static view. You don't need to worry about which type of page
   you're visiting.
   """
-  def visit(conn, path) do
-    case get(conn, path) do
-      %{assigns: %{live_module: _}} = conn ->
-        PhoenixTest.Live.build(conn)
-
-      %{status: 302} = conn ->
-        path = redirected_to(conn)
-
-        conn
-        |> recycle(all_headers(conn))
-        |> visit(path)
-
-      conn ->
-        PhoenixTest.Static.build(conn)
-    end
+  def visit(%Plug.Conn{} = conn, path) do
+    ConnHandler.visit(conn, path)
   end
 
-  defp all_headers(conn) do
-    Enum.map(conn.req_headers, &elem(&1, 0))
+  def visit(initial_struct, path) do
+    Driver.visit(initial_struct, path)
   end
 
   @doc """
