+++ lib/phoenix_test/driver.ex
@@ -1,5 +1,6 @@
 defprotocol PhoenixTest.Driver do
   @moduledoc false
+  def visit(initial_struct, path)
   def render_page_title(session)
   def render_html(session)
   def click_link(session, selector, text)
